+++ packages/yourloops/pages/user-account/sections/data-sharing-section/data-sharing-section.tsx
@@ -52,7 +52,7 @@ export const DataSharingSection: FC = () => {
   const fetchExternalConsents = useCallback(() => {
     setRefreshInProgress(true)
 
-    ExternalConsentsApi.getConsents(userId)
+    ExternalConsentsApi.getConsents()
       .then((consents: ExternalConsent[]) => {
         setConsents(consents)
         return consents