+++ test/locales/translations.test.js
@@ -337,7 +337,11 @@ const ignoredTransKeyForYourLoops = [
   'remove-clinician-question',
   'remove-clinician-question-patient',
   'revoke-consent-question',
-  'button-download-ifu'
+  'button-download-ifu',
+  'data-access-granted-description',
+  'data-access-denied-description',
+  'data-access-error-description',
+  'data-access-request-description1'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   // Documentation!