+++ CHANGELOG.md
@@ -4,15 +4,16 @@ It is based on Tidepool Blip 1.27.
 
 ## 3.10.0-beta - 2026-04-15
 ### Added
+- YLP-3892 Accept data sharing request from external partners
 - YLP-3893 Data sharing management with external partners
 - YLP-3904 As an HCP or a patient, I would like to be able to download the IFU corresponding to the system used, directly from Yourloops
 - YLP-3907 Add a conditional rendering on birth date display
 
+### Changed
+- YLP-3858 Allowing HCP to add an 11-digit phone number.
+
 ### Engineering
 - YLP-3885 Daily components migration (data elements only)
-
-### Fixed
-- YLP-3858 Allowing HCP to add an 11-digit phone number.
 
 ## 3.9.1 - 2026-05-04
 ### Fixed